refactor(All): Lotto 패키지 변경
