+++ src/main/java/lotto/domain/LottoMachine.java
@@ -3,7 +3,6 @@ package lotto.domain;
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
 import java.util.List;
-import lotto.Lotto;
 
 public class LottoMachine {
