+++ src/main/java/lotto/domain/WinningLotto.java
@@ -1,7 +1,6 @@
 package lotto.domain;
 
 import java.util.List;
-import lotto.Lotto;
 import lotto.exception.LottoErrorMessage;
 import lotto.exception.WinningLottoErrorMessage;
